refactor(client): 이미지 업로드 로직 공통화 및 폼 로직을 커스텀 훅으로 분리
