+++ apps/client/src/app/community/[id]/_components/comment-input.tsx
@@ -1,83 +1,24 @@
 "use client";
 
 import Image from "next/image";
-import { useRouter } from "next/navigation";
-import { type ChangeEvent, useEffect, useRef, useState } from "react";
-import { getPresignedUrlAction } from "@/common/api/actions/image.actions";
-import { getFileExtension, uploadToS3 } from "@/common/utils/upload-to-s3";
-import { postCommentAction } from "../../_actions/comment-actions";
+import { useCommentForm } from "../../_hooks/use-comment-form";
 import * as s from "./styles/comment-input.css";
 
-const MAX_IMAGE_SIZE_MB = 10;
-
 export const CommentInput = ({ boardId }: { boardId: number }) => {
-	const router = useRouter();
-	const [text, setText] = useState("");
-	const [image, setImage] = useState<File | null>(null);
-	const [imageError, setImageError] = useState<string | null>(null);
-	const [isSubmitting, setIsSubmitting] = useState(false);
-	const [submitError, setSubmitError] = useState<string | null>(null);
-	const fileInputRef = useRef<HTMLInputElement>(null);
-
-	const previewUrl = useImageObjectUrl(image);
-	const canSubmit = text.trim().length > 0;
-
-	const handleFileChange = (e: ChangeEvent<HTMLInputElement>) => {
-		const file = e.target.files?.[0];
-		if (!file) return;
-
-		if (file.size > MAX_IMAGE_SIZE_MB * 1024 * 1024) {
-			setImageError(`파일 크기는 ${MAX_IMAGE_SIZE_MB}MB를 초과할 수 없습니다.`);
-			e.target.value = "";
-			return;
-		}
-
-		setImageError(null);
-		setImage(file);
-	};
-
-	const removeImage = () => {
-		setImage(null);
-		setImageError(null);
-		if (fileInputRef.current) fileInputRef.current.value = "";
-	};
-
-	const handleSubmit = async () => {
-		if (!canSubmit || isSubmitting) return;
-		setIsSubmitting(true);
-
-		try {
-			let imageUrl: string | null = null;
-			if (image) {
-				const urlResult = await getPresignedUrlAction({
-					prefix: "board",
-					extension: getFileExtension(image),
-				});
-				if (!urlResult.success) throw new Error(urlResult.error);
-				await uploadToS3(urlResult.data.presignedUrl, image);
-				imageUrl = urlResult.data.publicUrl;
-			}
-
-			const result = await postCommentAction({
-				boardId,
-				payload: {
-					content: text,
-					...(imageUrl && { photoUrl: imageUrl }),
-				},
-			});
-
-			if (!result.success) throw new Error(result.error);
-
-			setText("");
-			removeImage();
-			setSubmitError(null);
-			router.refresh();
-		} catch (err) {
-			setSubmitError(err instanceof Error ? err.message : "오류가 발생했습니다.");
-		} finally {
-			setIsSubmitting(false);
-		}
-	};
+	const {
+		text,
+		setText,
+		image,
+		previewUrl,
+		imageError,
+		submitError,
+		canSubmit,
+		isSubmitting,
+		fileInputRef,
+		handleFileChange,
+		removeImage,
+		handleSubmit,
+	} = useCommentForm(boardId);
 
 	return (
 		<div className={s.wrapper}>
@@ -138,24 +79,4 @@ export const CommentInput = ({ boardId }: { boardId: number }) => {
 			{submitError && <p className={s.warningText}>{submitError}</p>}
 		</div>
 	);
-};
-
-const useImageObjectUrl = (file: File | null) => {
-	const [url, setUrl] = useState<string | null>(null);
-
-	useEffect(() => {
-		if (!file) {
-			setUrl(null);
-			return;
-		}
-
-		const objectUrl = URL.createObjectURL(file);
-		setUrl(objectUrl);
-
-		return () => {
-			URL.revokeObjectURL(objectUrl);
-		};
-	}, [file]);
-
-	return url;
 };
