+++ apps/client/src/common/hooks/use-image-object-url.ts
@@ -1,0 +1,19 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+export const useImageObjectUrl = (file: File | null): string | null => {
+	const [url, setUrl] = useState<string | null>(null);
+
+	useEffect(() => {
+		if (!file) {
+			setUrl(null);
+			return;
+		}
+		const objectUrl = URL.createObjectURL(file);
+		setUrl(objectUrl);
+		return () => URL.revokeObjectURL(objectUrl);
+	}, [file]);
+
+	return url;
+};
